解决梯度问题，添加测试集模块，目前Correctness约为95%

+++ dnn/fclayer/fcLayer.hpp
@@ -72,7 +72,7 @@ void fcLayer::sigmoid()
 	{
 		float tmp = neurals.at(i).getOutput();
 
-		tmp = 1 / (1 + exp(-tmp));
+		tmp = 1 / (1 + expf(-tmp));
 		neurals.at(i).setOutput(tmp);
 	}
 }
